email_system/responder/requirements.txt
kafka-python
jinja2